requirements.txt
aiofiles==23.2.1
aiogram==3.3.0
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
attrs==23.2.0
build==1.2.1
CacheControl==0.14.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cleo==2.1.0
crashtest==0.4.1
cryptography==42.0.7
distlib==0.3.8
dnspython==2.5.0
dulwich==0.21.7
fastjsonschema==2.19.1
filelock==3.14.0
frozenlist==1.4.1
idna==3.6
importlib_metadata==7.1.0
installer==0.7.0
jaraco.classes==3.4.0
jeepney==0.8.0
keyring==24.3.1
magic-filter==1.0.12
more-itertools==10.2.0
motor==3.3.2
msgpack==1.0.8
multidict==6.0.4
packaging==24.0
pexpect==4.9.0
pip==24.0
pkginfo==1.10.0
platformdirs==4.2.1
poetry==1.8.3
poetry-core==1.9.0
poetry-plugin-export==1.8.0
ptyprocess==0.7.0
pycparser==2.22
pydantic==2.5.3
pydantic_core==2.14.6
pydantic-settings==2.1.0
pymongo==4.6.1
pyproject_hooks==1.1.0
python-decouple==3.8
python-dotenv==1.0.1
rapidfuzz==3.9.0
requests==2.31.0
requests-toolbelt==1.0.0
SecretStorage==3.3.3
setuptools==69.5.1
shellingham==1.5.4
tomlkit==0.12.5
trove-classifiers==2024.4.10
typing_extensions==4.9.0
urllib3==2.2.1
virtualenv==20.26.1
wheel==0.43.0
yarl==1.9.4
zipp==3.18.1
